requirements.txt
asgiref==3.4.1
betterpath==0.2.2
boto3==1.23.10
botocore==1.26.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
d==0.2.2
dataclasses==0.8
dj-database-url==1.3.0
Django==3.2.18
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.6
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
django-ipware==4.0.2
django-project==0.7
django-session-timeout==0.1.0
django-storages==1.12.3
django-widget-tweaks==1.4.11
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.8.3
jmespath==0.10.0
l==0.11.0
lxml==4.9.2
Markdown==3.3.7
newsapi-python==0.2.7
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.21
Pygments==2.14.0
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
typing_extensions==4.1.1
urllib3==1.26.15
vcversioner==2.16.0.0
whitenoise==5.3.0
zipp==3.6.0
zope.interface==5.5.2
